Add explicit int cast for DEBUG_BOOL hack (#6325)
git-svn-id: https://svn.osgeo.org/gdal/trunk@33099 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/netcdf/netcdfdataset.cpp
@@ -5773,7 +5773,7 @@ int netCDFDataset::DefVarDeflate(
         // TODO: make sure this is okay.
         CPLDebug( "GDAL_netCDF",
                   "DefVarDeflate( %d, %d ) nZlevel=%d",
-                  nVarId, bChunkingArg, nZLevel );
+                  nVarId, static_cast<int>(bChunkingArg), nZLevel );
 
         int status = nc_def_var_deflate(cdfid,nVarId,1,1,nZLevel);
         NCDF_ERR(status);
